fix(delta-lake): fix dependency issue for snowflake due to s3_util
