+++ metadata-ingestion/src/datahub/ingestion/source/s3/source.py
@@ -51,12 +51,11 @@ from datahub.ingestion.api.decorators import (
 )
 from datahub.ingestion.api.source import Source, SourceReport
 from datahub.ingestion.api.workunit import MetadataWorkUnit
+from datahub.ingestion.source.aws.s3_boto_utils import get_s3_tags, list_folders
 from datahub.ingestion.source.aws.s3_util import (
     get_bucket_name,
     get_bucket_relative_path,
     get_key_prefix,
-    get_s3_tags,
-    list_folders,
     strip_s3_prefix,
 )
 from datahub.ingestion.source.data_lake.data_lake_utils import ContainerWUCreator
